Nate Woods Coco/Image tool update: This patch:
Core:
- Adds new endian macros:
     BIG_ENDIANIZE_SHORT
     BIG_ENDIANIZE_LONG
     LITTLE_ENDIANIZE_SHORT
     LITTLE_ENDIANIZE_LONG

I propose that these macros become standard because it seems that
there are tons of drivers that have their own private endian macros,
which is silly.  No rush to immediately convert everything over, but
as time goes on it would be nice to standardize.

CoCo/Dragon:
- Fixes a bug the CoCo 3 horizontal scroll so Crystal City is
partially playable now
- Converts the CoCo/Dragon driver to use an include file
- Moves PAK loading code into a new src/mess/formats directory

Imgtool:
- Adds a CRC command to imgtool so output image information in .crc
format (as per Stefano)
- basename()'s the command name when printing error info
- Fixed a typo in the usage examples
- Adds a very rough imgtool.txt to document imgtool