Phil Stroffolino <pstroffo@yahoo.com> wrote: >- there are priority bugs (the usual sprite-tilemap >orthogonality).  According to Andrew Prime, Fantasy >Zone is a good place to see them.
I have implemented in the Deniam driver (though the sprite drawing routine
is a quick hack), I thought it was already supported in sprite.c.

>- there are special codes in sprite pen data, marking
>left and right edges of sprites. If an "end of data"
>marker is encountered while the sprite hardware is
>blitting a sprite, the next line of sprite is drawn,

I am problably doing something wrong, because from how the Deniam games use
sprites, it seems it waits for the mark also before starting drawing. This
isn't consistent with the ESWAT title screen so it must be wrong.

>- we need a general purpose "streamed sprite" blitter,
>that handles variable-sized sprites.

I guess so. There already are low level pixel-stuffing functions in
drawgfx.c, but the end of line markers complicate things

>Thierry, can you confirm confirm whether the screen
>blanks or freezes on a real PCB?

On second thought, there are long pauses in logicpro with the screen
disabled, and they look ugly if it's blanked, so probably it really freezes.

I'm attaching the Deniam driver because there is an useful testbed fpr the
tilemap optimizations you promised several months ago.
Before every level, logicpro shows a screen with some random stuff in it
(including a yellow submarine), and scrolls it around at random at a furious
pace. This literally brings the tilemap code to its knees, raising frameskip
to 11 and still not being able to cope with it on my machine.

BTW, in case you hadn't noticed, the profiler now has three entries
dedicated to the tilemap code (update, render and draw). Render definitely
needs to be improved, it sometimes has peaks of 70% or so.

Nicola Salmoria