Basically, the system is a Z80 BIOS with a built in NES, and 10 cart slots. There were two different layouts, single monitor and dual monitors. I emulate dual-monitors, but I commented how the system work in single monitor mode. I ripped off Brad's code from MESS to base the PPU emulation, but I went a step further and virtualized it completely, so you can have multiple PPU's running concurrently. Hopefully this will help getting the Nintendo VS DualSystem running.
Now, each cartridge has a security serial prom, so your favorite local
operator couldnt just buy a new game from the local Toy R' Us, and put
it into a cart.
We have most game dumps, but we're missing a lot of these proms.
The ones we currently have, were manually read by Al (thank you).
However, I just wrote a ROM for the BIOS that can dump the contents
of the prom to the screen, so if you are a PC10 board owner, and
have any of the carts with the missing prom data, mail me up and
I'll send you this ROM.

Note I only mapped the games for wich we do have security prom
dumps. It's pointless to map the rest, as they wont even start up.

Take a look at the driver's top to see a list of games, what's working,
what's not, and why.

Ernesto.