I've updated the 68k core with fixes for compiles without 68010 and the NULL fix, as well as a small fix to 68020 throwaway stack frames (used when an interrupt occurs when the CPU is in master mode).
I can't find the problem that Nicola was mentioning with sf2accp2... The
code seems to be running fine, but it appears to be waiting for another
device.
If he could give me some more information about the problem, maybe I can
track it down easier.  Was it working in a previous version of MAME?
I did notice that they modify stack frames and then fake RTE's and RTS's
like there's no tomorrow.

Karl Stenerud