This is an attempt to implement the starfield background in Strider and Forgotten Worlds.
It is completely wrong, but the scrolling is right ;-)

I am quite sure that the star placement is taken from the beginning of the
gfx ROMs (the first 0x4000 bytes don't decode as normal tiles, and are the
same in strider and forgottn). It's not clear, however, how the mapping
should be done.

Some screen shots of the Forgotten Worlds title screen might be helpful,
since stars are colored and don't scroll. In Strider they are only shades of
grey and it's difficult to pick a place where they don't scroll while giving
a good view of a large portion of the background.


Nicola Salmoria