bjtwin?
It is again a bug in the 68000 core. The line F emulation exception must
store in the stack the address of the offending instruction, not of the
following one. Therefore, in m68kcpu.h/m68ki_exception_1111(), replace the
two references to REG_PC with REG_PC-2.

This does NOT fix sf2accp2 which is still broken.

Nicola Salmoria