Here's a first try at that vicdual driver sound hardware. The schematics are two pages full of always repeating but always slightly different hardware - and the heart-beat section is what I yet have to understand, so this is not complete.
Should I add the various parts of the sound as distinct channels,
so they could be balanced using the volume control settings?
There are 12 types of sounds (11 of which are emulated now)
and currently they're mixed in a way that 4 at the same time
won't clip. But mixing them so that all voices are equally
prominent seems difficult, due the fact that you perceive a
1-2kHz tone much louder than a 190Hz hum :-P I'm not sure
if mixing 12 channels in streams.c would be a good idea?

Adding the CUSTOM_SOUND to the driver could probably be done
in a nicer way too.

Juergen