"Vasvari, Zsolt" <Zsolt.Vasvari@ca.com> wrote: > Terribly sorry, attached again.  MAME32 is still on B5, I will be more > careful in the future.
Here is Commando again, simplified and with VIDEO_BUFFERS_SPRITERAM instead
of the vblank hack, since I think we agreed this is how the Capcom games
work. I have to say that I studied the Commando schematics and I don't see
the buffering happening, but they are a little over my capabilities. Anyway,
this keeps sprites in sync but of course does nothing for the disappearing
sprites - which might just as well be in the original.

About the question that you put in the source: yes, vertical scroll is used
by the game, during intermissions.

Note that you get a black background when you toggle the flip screen dip
switch. This is a bug in the tilemap system that I already reported some
time ago: the tilemap scroll position is not updated when tilemap_set_flip()
is called. Since the game doesn't write to the vertical scroll register
every frame, the scroll position is not updated until the first
intermission.
I'm not putting a hack in the driver to work around that because this is a
bug in tilemap.c and must be fixed there.

> There is one more remaining issue with Vulgus, namely the redraw artifacts
> visible between screen changes.  I can't find a video off register/bit or
> anything.

This was the first game made by Capcom, it might just be that way. Though it
does look ugly.

Nicola Salmoria