Fantasia and New Fantasia are almost ROM swaps too.
There is a ROM in the newfant set, obj2_14.rom, which is identical to t2.107
from the Terminator 2 set. I'm not using it and it doesn't seem to be
needed.
There are four unused ROMs in the fantasia set too, again the game seems to
work fine without them.

Nicola Salmoria