Thanks to Aaron's very useful comments, here is an update with proper timming (finally). The side effect is that there's no more tearing on scrolling games.
I also fixed Ninja Gaiden 3 while I was at it.

Brad: You might want to take a look at Ninja Gaiden 3. On the intro, after it says 'Tecmo
presents', there's a ninja jumping, and somehow I think it's wrong in MESS. The screen is
setup to look like 'widescreen', with black bars on the top and the bottom. On this PC10
driver you can see the sprite getting 'clipped' by this black stripe (I think it's the right
behaviour), while on Mess it's on top of it. It might have to do with any recent changes you
made, since my PPU code is entirely based on your PPU core circa beta 3 or 4.
FWIW, you can even try the PC10 rom. It uses a MMC3 mapper (nr 4).

Howie: To use this updated PPU code on the VS driver, you need to change the init
functions on machine/vsnes.c.
There change the reset function for the ppu to look like this:
For the unisystem games:
    ppu2c03b_reset( 0, 1 );

For the dualsystem games:
    ppu2c03b_reset( 0, 1 );
    ppu2c03b_reset( 1, 1 );

Ernesto.