Enclosed is the current driver for the Atari polygon games. Only STUN Runner is fully playable with sound, but that's good enough for the next beta IMO.
It's probably okay to leave Hard Drivin' hooked up. It runs through attract
mode and mostly kinda plays, but there appears to be a nasty 34010 bug that
causes the physics to go out of control after a crash. The gear shift isn't
hooked up, and there's no sound. If you start it up, you'll be asked to
calibrate a bunch of stuff. The "key" is button 2. Player 1's pedal is the
accelerator, player 2's pedal is the brake, player 3's pedal is the clutch.

The remaining games are definitely TESTDRIVER material. Race Drivin' Compact
will boot into attract mode, but the two ASICs aren't emulated so it goes
wonky pretty quickly in the simulation. Race Drivin' Upgrade almost boots,
but I think I have an incomplete dump and the ROMs I borrowed from the
Compact version don't match up completely.

Note that I've included an updated slapstic.c with info for the Race Drivin'
slapstic.

Steel Talons will start to boot, but it uses a more advanced ADSP board that
I'm still working on.

Aaron