> unemulated/Alpha/SkyAdventure-Alpha.zip > > this is the "Original" set. Works in the current driver, but can not be > coined-up Does if you hit the service key :)  None the less, here's a real fixed driver (with added flipscreen), plus the text layer fixed on Super Baseball, no idea how long that had been broken. I assume this set (Alpha Denshi) can be called the original Japanese version?
Also included:  Pushman, 1990 Comad/Sammy

A nice block pushing game, in keeping with Comad's traditions there's a bare breasted woman
in the tile graphics (2 bpp!), though it doesn't appear to be used anywhere.

It uses a 68705, which I'm no expert on, I notice all the other 68705 games have 2k of ROM,
and the core reflects this - well this has 4k and therefore the start vector is in a different
place.  I don't know if it's a special 68705 or what, so I'm leaving that for someone else to figure
out.  You only need to change 2 things in the core to get it to work (6805.c):

Line 325:  Change       PC |= 0xf800; to        PC |= 0xf000;

Make m68705_reset look like this:

m68705_reset(void *param)
{
    m6805_reset(param);
        /* Overide default 6805 type */
        m6805.subtype = SUBTYPE_M68705;
        AMASK   = 0xfff;
        RM16( 0xffe, &m6805.pc );
}

Bryan