"Aaron Giles" <2giles@home.com> wrote: > I don't know who changed this code in memory.c or why it was changed
I sent a message to the list on 18/8 (subj: "memory bugfix"), with the
changed function and explanation.
Your code was broken because for 1-byte long memory ranges, it did a
memset(start,0,-1) causing pyrotechnic crashes.

To fix it properly, take the CURRENT code (not the b5 one) and replace (note
the wrong comment that was there):

    /* find the base of the lowest memory region that extends past the end
*/
    for (mra = Machine->drv->cpu[cpu].memory_read; mra->start != -1; mra++)
     if (mra->start <= end && mra->end > end) end = mra->end;
    for (mwa = Machine->drv->cpu[cpu].memory_write; mwa->start != -1; mwa++)
     if (mwa->start <= end && mwa->end > end) end = mwa->end;

with

    /* find the end of the contiguous block of memory */
    for (mra = Machine->drv->cpu[cpu].memory_read; mra->start != -1; mra++)
     if (mra->start <= end+1 && mra->end > end) end = mra->end;
    for (mwa = Machine->drv->cpu[cpu].memory_write; mwa->start != -1; mwa++)
     if (mwa->start <= end+1 && mwa->end > end) end = mwa->end;


Nicola Salmoria