"Bryan McPhail" <mish@mame.net> wrote: > Bomberman World/Atomic Punk working except for some cosmetic glitches
There seems to be a uninitialized variable bug, the glitches are not the
same on every run. E.g. sometimes on startup the background behind the title
screen is half black.

> no sound yet, but it should be the same as the M72 stuff.

Here it is working, ADPCM samples are not played. The NMI handler is just
RETN, so like some M72 games this must have external hardware taking care of
the samples.

Nicola Salmoria