"Yasuhiro Ogawa" <yasuoga@tvt.ne.jp> wrote: >  Nouryoku Koujou Iinkai (c)1995 Tecmo >  Similar (same?) hardware to bjtwin.
Indeed.
Sound is completely wrong, I think, can you confirm?
Also, there is probably a bug in the video driver, as you can see in the
attached shot,the background is 0123456789ABCDEF instead of graphics. This
however might be how the game is so I'd like to have confirmation before
investigating.


Nicola Salmoria