This also adds Hacha Mecha Fighter, but all it does is boot to the tile screen. The (custom?) sound cpu is used for a more complex protection then macross: it shares some RAM, and the main CPU jumps to subroutines there.
However there are only 16 subroutines, and they are all no more than 16
bytes long, so it's possible they are just jump vectors, which I think can
assume two different values since the main CPU can initialize the RAM in two
different ways.
With some patience it might be possible to work out what they are supposed
to point to.

With an ICE it should be trivial to get it going.

Nicola Salmoria